Begin adding button functionality
There is still some kinks to work out with the functionality of the
grid button. However the logic has been implemented that enables it to
prompt the user for a new number limited to between 1 and 100, that will
generate a new grid of square divs for both rows and columns with the
number entered.